+++ SKCSwift/Models/HeartAPI.swift
@@ -22,3 +22,20 @@ struct YouTubeVideos: Codable, Equatable {
     let thumbnailUrl: String
     let url: String
 }
+
+struct Events: Codable {
+    let service: String
+    let events: [Event]
+}
+
+struct Event: Codable, Equatable {
+    static func == (lhs: Event, rhs: Event) -> Bool {
+        lhs.name == rhs.name && lhs.eventDate == rhs.eventDate && lhs.url == rhs.url
+    }
+    
+    let name: String
+    let notes: String
+    let location: String
+    let eventDate: String
+    let url: String
+}
