--- SKCSwift/Models/SearchResults.swift
@@ -1,14 +1,0 @@
-//
-//  SearchResults.swift
-//  SKCSwift
-//
-//  Created by Javi Gomez on 1/8/23.
-//
-
-import Foundation
-
-struct SearchResults: Identifiable, Equatable {
-    let id = UUID()
-    let section: String
-    let results: [Card]
-}
