--- SKCSwift/Models/Trending.swift
@@ -1,17 +1,0 @@
-//
-//  Trending.swift
-//  SKCSwift
-//
-//  Created by Javi Gomez on 7/22/24.
-//
-
-struct TrendingMetric<R:Codable>: Codable {
-    let resource: R
-    let occurrences: Int
-    let change: Int
-}
-
-struct Trending<R:Codable>: Codable {
-    let resourceName: TrendingResourceType
-    let metrics: [TrendingMetric<R>]
-}
